loading moved to a separate thread

+++ Source/Components/SampleNavigator.cpp
@@ -95,7 +95,7 @@ void SampleNavigator::paintOverChildren(juce::Graphics& g)
                     auto pos = jmap<float>(float(location), 0.f, float(sample->getNumSamples()), 0.f, float(painter.getWidth()));
 
                     Path voicePath{};
-                    voicePath.addLineSegment(Line<int>(pos, 0, pos, getHeight()).toFloat(), 1.f);
+                    voicePath.addLineSegment(Line<float>(pos, 0, pos, getHeight()), 1.f);
                     g.setColour(lnf.VOICE_POSITION_COLOR.withAlpha(voice->getEnvelopeGain()));
                     g.strokePath(voicePath, PathStrokeType(1.f));
                 }
